[Refactor] PostView관련 소스 파일 내 기능들 refactor
