+++ travelPlan/Source/Feature/Feed/Constant/PostViewHeaderCategoryViewConstant.swift
@@ -1,0 +1,22 @@
+//
+//  PostViewHeaderCategoryViewConstant.swift
+//  travelPlan
+//
+//  Created by 양승현 on 2023/07/04.
+//
+
+import Foundation
+
+extension PostViewHeaderCategoryView {
+  struct Constant {
+    enum TotalView {
+      static let spacing: UISpacing = .init(
+        leading: 21.5, top: 10, bottom: 10)
+    }
+    
+    enum SortingView {
+      static let spacing: UISpacing = .init(
+        leading: 15, top: 10, trailing: 224.5, bottom: 10)
+    }
+  }
+}
