+++ travelPlan/Source/Feature/Feed/Protocol/PostView/MoreCategoryViewDelegate.swift
@@ -1,0 +1,10 @@
+//
+//  MoreCategoryViewDelegate.swift
+//  travelPlan
+//
+//  Created by 양승현 on 2023/07/04.
+//
+
+protocol MoreCategoryViewDelegate: AnyObject {
+  func didTapMoreCategoryView(with type: MoreCategoryType)
+}
